+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - None.
 ### Changed
-- None.
+- Moved `baseUrl` from `Endpoint` to `ApiProvider`.
 ### Deprecated
 - None.
 ### Removed